requirements.txt
flask
openai
praat-parselmouth
